requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
asn1crypto==1.5.1
atlassian-python-api==4.0.7
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==41.0.7
Deprecated==1.2.18
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
filelock==3.19.1
h11==0.16.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
packaging==25.0
platformdirs==3.11.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.10.1
pyOpenSSL==23.3.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.6.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.23
starlette==0.27.0
tomlkit==0.13.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
